LPD-33410 Reuse some fn on playwright
